% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{parse.datetime.flexible}
\alias{parse.datetime.flexible}
\title{Parse Datetime with Multiple Format Support}
\usage{
parse.datetime.flexible(datetime.string)
}
\arguments{
\item{datetime.string}{Character vector of datetime strings}
}
\value{
POSIXct datetime object
}
\description{
Internal function to parse datetime strings with automatic format detection.
}
\keyword{internal}
