% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{standardize.datetime}
\alias{standardize.datetime}
\title{Standardize Datetime Format}
\usage{
standardize.datetime(datetime.string)
}
\arguments{
\item{datetime.string}{Character vector of datetime strings}
}
\value{
Character vector in format "MM/DD/YYYY HH:MM:SS AM/PM"
}
\description{
Convert any parseable datetime to package standard format.
}
\examples{
standardize.datetime("10/7/2025 22:45")

}
