% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datetime_utils.R
\name{validate.sleep.data}
\alias{validate.sleep.data}
\title{Validate Sleep Data Format}
\usage{
validate.sleep.data(sleep.data, verbose = TRUE)
}
\arguments{
\item{sleep.data}{Data frame to validate}

\item{verbose}{Logical; if TRUE, print detailed diagnostics}
}
\value{
Logical; TRUE if valid, FALSE otherwise
}
\description{
Check if sleep data has correct columns and parseable datetime formats.
}
\examples{
data(sample.data)
validate.sleep.data(sample.data)

}
