% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.subject.log.R
\name{create.subject.log}
\alias{create.subject.log}
\title{Creating Subject Log from Raw Sleep Data}
\usage{
create.subject.log(sleep.data, output.type = "sleep")
}
\arguments{
\item{sleep.data}{Data frame with columns 'Subject Name', 'In Bed Time',
and 'Out Bed Time'. Multiple datetime formats supported.}

\item{output.type}{Character string: "sleep" for sleep periods (default)
or "wake" for wake periods}
}
\value{
Data frame with columns 'Subject Name', 'On Date', 'On Time',
'Off Date', 'Off Time', and 'Category'
}
\description{
Convert batch export format to subject log diary format by splitting datetime
into separate date and time columns. Optionally inverts sleep periods to wake
periods in a single step.
}
\examples{
data(sample.data)
sleep.log <- create.subject.log(sample.data, output.type = "sleep")
wake.log <- create.subject.log(sample.data, output.type = "wake")

}
